moved models to journal
