+++ libs/shared/modules/src/index.ts
@@ -3,5 +3,5 @@ export * from './application/application.module'
 export * from './case/case.module'
 export * from './health/health.module'
 export * from './journal'
-export * from './models'
+export * from './journal/models'
 export * from './util'
