+++ libs/shared/modules/src/journal/models/Advert.ts
