+++ libs/shared/modules/src/journal/models/AdvertAttachments.ts
