+++ libs/shared/modules/src/journal/models/AdvertDepartment.ts
