+++ libs/shared/modules/src/journal/models/index.ts
