+++ libs/shared/modules/src/util/advert-category-migrate.ts
@@ -1,6 +1,6 @@
 import { Category } from '@dmr.is/shared/dto'
 
-import { AdvertCategoryDTO } from '../models/AdvertCategory'
+import { AdvertCategoryDTO } from '../journal/models/AdvertCategory'
 import { advertMainCategoryMigrate } from './advert-main-category-migrate'
 
 export function advertCategoryMigrate(model: AdvertCategoryDTO): Category {
